Move edge line code out of editor to EdgeLine class
