+++ spyder/widgets/sourcecode/codeeditor.py
@@ -59,6 +59,7 @@ from spyder.widgets.arraybuilder import SHORTCUT_INLINE, SHORTCUT_TABLE
 from spyder.widgets.editortools import PythonCFM
 from spyder.widgets.sourcecode.base import TextEditBaseWidget
 from spyder.widgets.sourcecode.kill_ring import QtKillRing
+from spyder.widgets.panels.edgeline import EdgeLine
 
 try:
     import nbformat as nbformat
@@ -255,22 +256,6 @@ class ScrollFlagArea(QWidget):
         self.code_editor.wheelEvent(event)
 
 
-class EdgeLine(QWidget):
-    """Source code editor's edge line (default: 79 columns, PEP8)"""
-    def __init__(self, editor):
-        QWidget.__init__(self, editor)
-        self.code_editor = editor
-        self.column = 79
-        self.setAttribute(Qt.WA_TransparentForMouseEvents)
-
-    def paintEvent(self, event):
-        """Override Qt method"""
-        painter = QPainter(self)
-        color = QColor(Qt.darkGray)
-        color.setAlphaF(.5)
-        painter.fillRect(event.rect(), color)
-
-
 #===============================================================================
 # CodeEditor widget
 #===============================================================================
@@ -382,7 +367,6 @@ class CodeEditor(TextEditBaseWidget):
         self.setCursorWidth( CONF.get('main', 'cursor/width') )
 
         # 79-col edge line
-        self.edge_line_enabled = True
         self.edge_line = EdgeLine(self)
 
         # Blanks enabled
@@ -681,8 +665,8 @@ class CodeEditor(TextEditBaseWidget):
         self.set_scrollflagarea_enabled(scrollflagarea)
 
         # Edge line
-        self.set_edge_line_enabled(edge_line)
-        self.set_edge_line_column(edge_line_column)
+        self.edge_line.set_enabled(edge_line)
+        self.edge_line.set_column(edge_line_column)
 
         # Blanks
         self.set_blanks_enabled(show_blanks)
@@ -1359,17 +1343,6 @@ class CodeEditor(TextEditBaseWidget):
         """Trigger a calltip"""
         self.sig_show_object_info.emit(position)
 
-    #-----edge line
-    def set_edge_line_enabled(self, state):
-        """Toggle edge line visibility"""
-        self.edge_line_enabled = state
-        self.edge_line.setVisible(state)
-
-    def set_edge_line_column(self, column):
-        """Set edge line column value"""
-        self.edge_line.column = column
-        self.edge_line.update()
-    
     # -----blank spaces
     def set_blanks_enabled(self, state):
         """Toggle blanks visibility"""
